update id type to Long. closes #17

+++ src/main/java/com/rs/uspotus/repository/PresidentsRepository.java
@@ -3,5 +3,5 @@ package com.rs.uspotus.repository;
 import com.rs.uspotus.model.President;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PresidentsRepository extends CrudRepository<President, Integer> {
+public interface PresidentsRepository extends CrudRepository<President, Long> {
 }
